chore: update debug-logs
