+++ tests/spec/test-data/debug-logs/for-debug-logs.ts
@@ -1,0 +1,12 @@
+import { Router } from "@the/router";
+import { Just } from "maybe";
+import { Service } from "./service";
+export class ExampleComponent {
+    constructor(service: Service, router: Router, just: Just) {}
+
+    myMethod() {}
+
+    yourMethod() {}
+
+    theirMethod() {}
+}
